requirements.txt
absl-py==2.1.0
annotated-types==0.6.0
anyio==4.3.0
astunparse==1.6.3
CacheControl==0.14.0
cachetools==5.3.3
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
contourpy==1.2.1
cryptography==42.0.5
cycler==0.12.1
exceptiongroup==1.2.1
fastapi==0.110.2
firebase-admin==6.5.0
flatbuffers==24.3.25
fonttools==4.51.0
gast==0.5.4
gcloud==0.17.0
google-api-core==2.18.0
google-api-python-client==2.127.0
google-auth==2.29.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.0.0
google-cloud-core==2.4.1
google-cloud-firestore==2.16.0
google-cloud-storage==2.16.0
google-crc32c==1.5.0
google-pasta==0.2.0
google-resumable-media==2.7.0
googleapis-common-protos==1.63.0
grpcio==1.62.2
grpcio-status==1.62.2
h11==0.14.0
h5py==3.11.0
httplib2==0.22.0
idna==3.7
jws==0.1.3
keras==2.14.0
kiwisolver==1.4.5
libclang==18.1.1
Markdown==3.6
MarkupSafe==2.1.5
matplotlib==3.8.4
ml-dtypes==0.2.0
msgpack==1.0.8
numpy==1.24.3
oauth2client==3.0.0
oauthlib==3.2.2
opt-einsum==3.3.0
packaging==24.0
pillow==10.3.0
proto-plus==1.23.0
protobuf==4.25.3
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycparser==2.22
pycryptodome==3.4.3
pydantic==2.7.0
pydantic_core==2.18.1
PyJWT==2.8.0
pyparsing==3.1.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-jwt==2.0.1
python-multipart==0.0.9
requests==2.31.0
requests-oauthlib==2.0.0
requests-toolbelt==0.7.0
rsa==4.9
six==1.16.0
sniffio==1.3.1
starlette==0.37.2
tensorboard==2.14.1
tensorboard-data-server==0.7.2
tensorflow==2.14.0
tensorflow-estimator==2.14.0
tensorflow-io-gcs-filesystem==0.36.0
tensorflow-serving-api==2.5.0
termcolor==2.4.0
typing_extensions==4.11.0
uritemplate==4.1.1
urllib3==2.2.1
uvicorn==0.29.0
Werkzeug==3.0.2
wrapt==1.14.1
